requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.8.0
google-genai>=0.8.0
chromadb>=0.4.0
python-dotenv>=1.0.0
chardet>=5.0.0
